requirements.txt
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
colorama==0.4.6
cryptography==46.0.7
distro==1.9.0
Django==6.0.3
openai==1.63.2
googleapis-common-protos==1.74.0
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
proto-plus==1.27.2
protobuf==5.29.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
requests==2.33.1
sniffio==1.3.1
sqlparse==0.5.5
tenacity==9.1.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
websockets==16.0
whitenoise==6.12.0
gunicorn
python-dotenv==1.0.1
google-genai